+++ src/routes/__init__.py
@@ -14,7 +14,6 @@ from .variant import init_app as init_variant_app
 from .scans import init_app as init_scans_app
 from .scan_triggers import init_app as init_scan_triggers_app
 from .config import init_app as init_config_app
-from .notifications import init_app as init_notifications_app
 from .settings import init_app as init_settings_app
 from .frontpage import init_app as init_front_app
 
@@ -33,7 +32,6 @@ def init_app(app):
     init_scan_triggers_app(app)
     init_bulk_refresh_app(app)
     init_config_app(app)
-    init_notifications_app(app)
     init_settings_app(app)
     # keep front endpoint at the end
     init_front_app(app)